+++ main/windows/index.js
@@ -26,7 +26,7 @@ const mainWindow = () => {
     }
   })
 
-  win.loadURL(windowURL('home'))
+  win.loadURL(windowURL('dashboard'))
   return win
 }
