+++ renderer/components/dashboard/DashboardHeader.js
