+++ renderer/components/dashboard/StripedProgress.js
